Fix shift overflow UB and stop widening restored file permissions
- Decoder.c: 1<<i overflows a signed int at i==31. A build with
  -fsanitize=undefined traps on every --decode; use 1u<<i.
- Printer.c: same class of overflow in redundancy<<28 at -r 9 and -r 10.
- Fileproc.c: a page carries a single attribute bit, so the old mapping
  restored every file as 0644 (or 0444) and could publish a backup that was
  0600 before. Create the output under umask 0077 and chmod it to 0600,
  covering --force output and a pre-existing output file as well.

Both suites pass, including under a -fsanitize=undefined -fsanitize-trap=all
build. The Linux-only branch in Fileproc.c was not compiled here.

+++ src/Printer.c
@@ -676,7 +676,7 @@ static void Printnextpage(t_printdata *print) {
   // Now the most important part - encode and draw data, group by group!
   for (i=0; i<nstring; i++) {
     // Prepare redundancy block.
-    cksum.addr=offset ^ (redundancy<<28);
+    cksum.addr=offset ^ ((uint32_t)redundancy<<28);
     memset(cksum.data,0xFF,NDATA);
     // Process data group.
     for (j=0; j<redundancy; j++) {
